requirements.txt
flask
flask-cors
torch
transformers
numpy
requests
nltk
textblob
sentencepiece
tiktoken
protobuf
pydantic>=2.0.0
oxforddictionaries
python-dotenv
gunicorn
